chore: extract button enums
