+++ packages/shared/components/Button.svelte
@@ -1,30 +1,10 @@
-<script lang="typescript" context="module">
-    export enum ButtonSize {
-        Large = 'lg',
-        Medium = 'md',
-        Small = 'sm',
-    }
-
-    export enum ButtonVariant {
-        Primary = 'primary',
-        Caution = 'caution',
-        Warning = 'warning',
-    }
-
-    export enum HTMLButtonType {
-        Button = 'button',
-        Submit = 'submit',
-        Reset = 'reset',
-    }
-</script>
-
 <script lang="typescript">
-    import { Icon, Spinner } from 'shared/components'
+    import { ButtonSize, ButtonVariant, HTMLButtonType, Icon, Spinner } from 'shared/components'
     import { onMount } from 'svelte'
     import { appSettings } from '@core/app'
     import { bindEvents } from '@lib/utils'
     import { Icon as IconEnum } from '@lib/auxiliary/icon'
-    import { Event } from '@lib/typings/events'
+    import type { Event } from '@lib/typings/events'
 
     export let autofocus: boolean = false
     export let classes: string = ''
