+++ packages/shared/components/ToastContainer.svelte
@@ -7,7 +7,9 @@
 
     let toasts
     $: toasts = $displayNotifications.map((notification) => ({
-        ...notification,
+        type: notification.type,
+        alert: notification.alert,
+        message: notification.message,
         actions: notification.actions.map((action, actionIndex) => ({
             ...action,
             onClick: () => action.callback(notification, actionIndex),
@@ -19,10 +21,16 @@
     <ul class="space-y-2">
         {#each toasts as toast}
             <li in:fade={{ duration: 100 }} out:fade={{ duration: 100 }}>
-                {#if toast?.alert}
-                    <Alert classes={toast.classes} type={toast.type} message={toast.message} />
+                {#if toast.alert}
+                    <Alert type={toast.type} message={toast.message} />
                 {:else}
-                    <Toast {...toast} />
+                    <Toast
+                        type={toast.type}
+                        message={toast.message}
+                        subMessage={toast.subMessage}
+                        progress={toast.progress}
+                        actions={toast.actions}
+                    />
                 {/if}
             </li>
         {/each}
