+++ packages/shared/components/atoms/MenuItem.svelte
@@ -1,7 +1,7 @@
 <script lang="typescript">
     import { Icon, Text } from 'shared/components'
 
-    export let icon: string
+    export let icon: string = ''
     export let iconProps: Record<string, unknown> = undefined
     export let title: string
     export let subtitle = ''
