+++ packages/shared/components/atoms/buttons/ExportStrongholdButton.svelte
@@ -3,8 +3,7 @@
     import { localize } from '@core/i18n'
     import { showAppNotification } from '@lib/notifications'
     import { openPopup } from '@lib/popup'
-    import { Button } from 'shared/components'
-    import { ButtonSize } from 'shared/components/Button.svelte'
+    import { Button, ButtonSize } from 'shared/components'
 
     export let isBusy = false
     export let message = ''
