+++ packages/shared/components/enums/button-size.enum.ts
@@ -1,0 +1,5 @@
+export enum ButtonSize {
+    Large = 'lg',
+    Medium = 'md',
+    Small = 'sm',
+}
