+++ packages/shared/components/enums/button-type.enum.ts
@@ -1,0 +1,5 @@
+export enum HTMLButtonType {
+    Button = 'button',
+    Submit = 'submit',
+    Reset = 'reset',
+}
