+++ packages/shared/components/enums/button-variant.enum.ts
@@ -1,0 +1,5 @@
+export enum ButtonVariant {
+    Primary = 'primary',
+    Caution = 'caution',
+    Warning = 'warning',
+}
