+++ packages/shared/components/enums/index.ts
@@ -1,2 +1,5 @@
+export * from './button-type.enum'
+export * from './button-variant.enum'
+export * from './button-size.enum'
 export * from './font-weight.enum'
 export * from './text-type.enum'
