+++ packages/shared/components/molecules/DateInputButton.svelte
@@ -3,7 +3,7 @@
     import { Text, DateTimePicker, Icon } from 'shared/components'
     import { appSettings } from '@core/app'
 
-    export let value: string
+    export let value: string = undefined
 
     let customDate: Date
     let anchor: HTMLElement
@@ -24,7 +24,7 @@
 <button
     bind:this={anchor}
     on:click={() => (canShowDateTimePicker = true)}
-    class="flex flex-row justify-between border border-solid border-gray-300 bg-white text-center rounded-xl px-2 py-1"
+    class="flex flex-row justify-between border border-solid border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-800 hover:border-gray-500 dark:hover:border-gray-700 text-center rounded-xl px-2 py-1"
 >
     <Icon width="20" height="20" classes="text-gray-500" icon="calendar" />
     <Text>{formattedDate}</Text>
@@ -34,7 +34,6 @@
         position="top"
         {anchor}
         mode="date"
-        initialDate={value}
         bind:value={customDate}
         on:cancel={handleExpirationTimeCancelClick}
         on:confirm={handleExpirationTimeConfirmClick}
