+++ packages/shared/components/popups/AddNodePopup.svelte
@@ -1,15 +1,12 @@
 <script lang="typescript">
-    import { Text, NodeConfigurationForm, Button } from 'shared/components'
+    import { Text, NodeConfigurationForm, Button, HTMLButtonType } from 'shared/components'
     import { localize } from '@core/i18n'
-    import { INode, INetwork, addNodeToClientOptions, editNodeInClientOptions } from '@core/network'
+    import { INode, addNodeToClientOptions, editNodeInClientOptions } from '@core/network'
     import { closePopup } from '@lib/popup'
     import { activeProfile } from '@core/profile'
     import { showAppNotification } from '@lib/notifications'
-    import { HTMLButtonType } from 'shared/components/Button.svelte'
 
     export let node: INode = { url: '', auth: { username: '', password: '', jwt: '' } }
-    export let nodes: INode[] = []
-    export let network: INetwork
     export let isEditingNode: boolean = false
     export let onSuccess: (..._: any[]) => void
 
@@ -53,8 +50,6 @@
         bind:this={nodeConfigurationForm}
         bind:node
         {isBusy}
-        {nodes}
-        {network}
         isDeveloperProfile={$activeProfile.isDeveloperProfile}
     />
     <div class="flex flex-row justify-between space-x-4 w-full">
