+++ packages/shared/routes/dashboard/settings/views/advanced/MigrateLedgerIndex.svelte
@@ -1,7 +1,6 @@
 <script lang="typescript">
-    import { Button, Text } from 'shared/components'
+    import { Button, Text, ButtonSize } from 'shared/components'
     import { localize } from '@core/i18n'
-    import { ButtonSize } from 'shared/components/Button.svelte'
 </script>
 
 <Text type="h4" classes="mb-3">{localize('views.settings.migrateLedgerIndex.title')}</Text>
