+++ packages/shared/routes/dashboard/settings/views/help/HelpSection.svelte
@@ -1,9 +1,8 @@
 <script lang="typescript">
-    import { Button, Text } from 'shared/components'
+    import { Button, Text, ButtonSize } from 'shared/components'
     import { localize } from '@core/i18n'
     import { Platform } from 'shared/lib/platform'
     import { HelpAndInfoRoute } from '@core/router'
-    import { ButtonSize } from 'shared/components/Button.svelte'
 
     export let route: HelpAndInfoRoute
     export let url: string
